backend/requirements.txt
# Deep Learning & Reinforcement Learning
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
gymnasium==0.29.1

# API & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Database
pymongo==4.6.0
motor==3.3.2

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.1.3

# Visualization & Monitoring
tensorboard==2.15.1
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1

# Testing
pytest==7.4.3
httpx==0.25.2
